+++ src/alservice/al.py
@@ -125,7 +125,7 @@ class AccountLinking(object):
     Handles account linking logic
     """
 
-    def __init__(self, db: ALdatabase, salt: str, email_sender_create_account: Email,
+    def __init__(self, trusted_keys: list, db: ALdatabase, salt: str, email_sender_create_account: Email,
                  email_sender_pin_recovery: Email = None, pin_verify: str = None,
                  pin_empty: bool = True):
         """
@@ -139,6 +139,8 @@ class AccountLinking(object):
         self.pin_verify = None
         if pin_verify is not None:
             self.pin_verify = re.compile(pin_verify)
+
+        self.trusted_keys = trusted_keys
 
         self.pin_empty = pin_empty
         """:type: str"""
@@ -370,7 +372,7 @@ class AccountLinking(object):
             self.db.verify_account(email_hash, pin_hash)
             token = AccountLinking.create_token(email_hash, self.salt)
             self.db.save_token_state(token, email_hash)
-            self.email_sender_create_account.send_mail(token, email)
+            self.email_sender_pin_recovery.send_mail(token, email)
         except Exception as error:
             LOGGER.exception("Unknown error while changing pin.")
             raise ALserviceAuthenticationError() from error
